requirements.txt
Django==4.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-4-jet==1.0.9
django-ckeditor==6.5.1
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-widget-tweaks==1.4.12
ipython==8.7.0
numpy==2.1.3
openpyxl==3.0.10
pandas==2.2.3
Pillow==11.0.0
unicode-tr==0.6.1
pdfkit==1.0.0
reportlab==4.4.4